backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pillow==10.1.0
httpx==0.25.2
pydantic[email]==2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0
ultralytics==8.0.227
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
opencv-python-headless==4.8.1.78
groq==0.4.2
scikit-learn==1.3.2
